requirements.txt
# =========================
# Web Framework
# =========================
fastapi
uvicorn

# =========================
# PDF Processing (text-based PDFs)
# =========================
pdfplumber

# =========================
# Embeddings & Vector Search
# =========================
sentence-transformers
faiss-cpu

# =========================
# Reranking (CrossEncoder)
# =========================
torch

# =========================
# Data Validation & Config
# =========================
pydantic
python-dotenv

# =========================
# Utilities (safe & common)
# =========================
tqdm

boto3
python-dotenv
